@@ -2,34 +2,37 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": "^7.1.3",
+        "php": "^7.4",
         "ext-iconv": "*",
+        "ext-json": "*",
         "amenadiel/jpgraph": "^3.6",
         "calderacc/geobasic": "^0.2.0",
         "calderacc/yourls-api-manager": "^0.2.0",
         "doctrine/doctrine-migrations-bundle": "^1.0",
         "dragonmantank/cron-expression": "^2.1",
+        "flagception/flagception-bundle": "^3.2",
         "fresh/doctrine-enum-bundle": "^6.2",
         "friendsofsymfony/elastica-bundle": "^5.0",
         "friendsofsymfony/jsrouting-bundle": "^2.2",
-        "geocoder-php/nominatim-provider": "^5.0",
+        "geocoder-php/nominatim-provider": "^5.2",
         "hwi/oauth-bundle": "^0.6.1",
         "jms/serializer-bundle": "^2.3",
         "jublonet/codebird-php": "^3.1",
         "knplabs/knp-menu-bundle": "^2.2",
         "league/csv": "^9.1",
         "league/geotools": "^0.8.0",
-        "maxh/php-nominatim": "^1.2",
         "nelmio/api-doc-bundle": "^3.1",
         "php-amqplib/rabbitmq-bundle": "^1.14",
         "php-curl-class/php-curl-class": "^8.0",
         "php-http/guzzle6-adapter": "^1.1",
         "php-http/httplug-bundle": "^1.9",
         "php-http/message": "^1.7",
+        "predis/predis": "^1.1",
         "presta/sitemap-bundle": "^1.5",
         "salavert/time-ago-in-words": "^1.7",
         "sensio/framework-extra-bundle": "^5.1",
         "sonata-project/admin-bundle": "^3.40",
+        "sonata-project/datagrid-bundle": "^3.2",
         "sonata-project/doctrine-orm-admin-bundle": "^3.6",
         "sonata-project/intl-bundle": "^2.5",
         "sonata-project/seo-bundle": "^2.5",
@@ -41,7 +44,6 @@
         "symfony/flex": "^1.0",
         "symfony/form": "^4.0",
         "symfony/framework-bundle": "^4.0",
-        "symfony/lts": "^4@dev",
         "symfony/monolog-bundle": "^3.1",
         "symfony/orm-pack": "*",
         "symfony/process": "^4.0",
@@ -50,16 +52,21 @@
         "symfony/swiftmailer-bundle": "^3.1",
         "symfony/validator": "^4.0",
         "symfony/web-link": "^4.0",
-        "symfony/webpack-encore-pack": "*",
-        "symfony/yaml": "^4.0"
+        "symfony/webpack-encore-bundle": "^1.7",
+        "symfony/yaml": "^4.0",
+        "whiteoctober/breadcrumbs-bundle": "^1.4",
+        "willdurand/geocoder-bundle": "^5.0",
+        "ext-simplexml": "*"
     },
     "require-dev": {
+        "phpunit/phpunit": "^7.0",
         "symfony/browser-kit": "^4.0",
         "symfony/css-selector": "^4.0",
         "symfony/dotenv": "^4.0",
         "symfony/maker-bundle": "^1.0",
-        "symfony/phpunit-bridge": "^4.0",
-        "symfony/profiler-pack": "*"
+        "symfony/phpunit-bridge": "^4.3",
+        "symfony/profiler-pack": "*",
+        "symfony/test-pack": "^1.0"
     },
     "config": {
         "preferred-install": {
