@@ -3,6 +3,7 @@
 namespace App\Geocoding\Guesser;
 
 use Caldera\GeoBasic\Coord\Coord;
+use Geocoder\Model\AddressCollection;
 use Geocoder\Query\ReverseQuery;
 
 class CityGuesser extends AbstractGuesser implements CityGuesserInterface
@@ -16,12 +17,15 @@ class CityGuesser extends AbstractGuesser implements CityGuesserInterface
 
     public function guess(Coord $coord): ?string
     {
-        $result = $this->geocoder->reverseQuery(ReverseQuery::fromCoordinates($coord->getLatitude(), $coord->getLongitude()));
+        $query = ReverseQuery::fromCoordinates($coord->getLatitude(), $coord->getLongitude());
 
-        if (!$result || !$result->first() || (!$result->first()->getLocality() && !$result->first()->getSubLocality())) {
+        /** @var AddressCollection $addressCollection */
+        $addressCollection = $this->provider->reverseQuery($query);
+
+        if (!$addressCollection || !$addressCollection->first() || (!$addressCollection->first()->getLocality() && !$addressCollection->first()->getSubLocality())) {
             return null;
         }
 
-        return $result->first()->getLocality() ?? $result->first()->getSubLocality();
+        return $addressCollection->first()->getLocality() ?? $addressCollection->first()->getSubLocality();
     }
 }
