@@ -2,13 +2,13 @@
 
 namespace App\Pollution\PollutionDataFactory;
 
+use App\Air\ViewModel\MeasurementViewModel;
 use App\Entity\Data;
-use App\Entity\Station;
-use App\Pollution\Box\Box;
+use App\Pollution\UniqueStrategy\Hasher;
 
 class PollutionDataFactory extends AbstractPollutionDataFactory
 {
-    public function createDecoratedPollutantList(\DateTime $dateTime = null, \DateInterval $dateInterval = null, int $workingSetSize = 1): array
+    public function createDecoratedPollutantList(\DateTime $dateTime = null, \DateInterval $dateInterval = null, int $workingSetSize = 20): array
     {
         if (!$dateTime) {
             $dateTime = new \DateTime();
@@ -22,11 +22,11 @@ class PollutionDataFactory extends AbstractPollutionDataFactory
 
         $dataList = $this->getDataListFromStationList($dateTime, $dateInterval, $workingSetSize);
 
-        $boxList = $this->getBoxListFromDataList($dataList);
+        $measurementViewModelList = $this->getMeasurementViewModelListFromDataList($dataList);
 
-        $boxList = $this->decoratePollutantList($boxList);
+        $measurementViewModelList = $this->decoratePollutantList($measurementViewModelList);
 
-        return $boxList;
+        return $measurementViewModelList;
     }
 
     protected function getDataListFromStationList(\DateTime $fromDateTime = null, \DateInterval $interval = null, int $workingSetSize): array
@@ -50,32 +50,32 @@ class PollutionDataFactory extends AbstractPollutionDataFactory
                 }
             }
         }
-
+        
         return $this->dataList->getList();
     }
 
-    protected function getBoxListFromDataList(array $dataList): array
+    protected function getMeasurementViewModelListFromDataList(array $dataList): array
     {
-        $pollutantList = [];
+        $measurementViewModelList = [];
 
         /** @var array $data */
         foreach ($dataList as $data) {
             /** @var Data $dataElement */
             foreach ($data as $dataElement) {
                 if ($dataElement) {
-                    $pollutantList[$dataElement->getPollutant()][$dataElement->getId()] = new Box($dataElement);
+                    $measurementViewModelList[$dataElement->getPollutant()][Hasher::hashData($dataElement)] = new MeasurementViewModel($dataElement);
                 }
             }
         }
 
-        return $pollutantList;
+        return $measurementViewModelList;
     }
 
     protected function decoratePollutantList(array $pollutantList): array
     {
         return $this
             ->reset()
-            ->boxDecorator
+            ->measurementViewModelFactory
             ->setCoord($this->coord)
             ->setPollutantList($pollutantList)
             ->decorate()
