@@ -8,7 +8,7 @@ use Caldera\GeoBasic\Coord\CoordInterface;
 
 interface PollutionDataFactoryInterface
 {
-    public function createDecoratedPollutantList(\DateTime $dateTime = null, \DateInterval $dateInterval = null, int $workingSetSize = 1): array;
+    public function createDecoratedPollutantList(\DateTime $dateTime = null, \DateInterval $dateInterval = null, int $workingSetSize = 20): array;
     public function setCoord(CoordInterface $coord): PollutionDataFactoryInterface;
     public function setStation(Station $station): PollutionDataFactoryInterface;
     public function setStrategy(PollutantFactoryStrategyInterface $strategy): PollutionDataFactoryInterface;
