@@ -2,46 +2,63 @@
 
 namespace App\Provider\UmweltbundesamtDe\StationLoader;
 
+use App\DBAL\Types\AreaType;
+use App\DBAL\Types\StationType;
 use App\Entity\Station;
 use App\Provider\AbstractStationLoader;
 use App\Provider\StationLoaderInterface;
 use App\Provider\UmweltbundesamtDe\UmweltbundesamtDeProvider;
 use Curl\Curl;
 use Doctrine\ORM\EntityManager;
-use League\Csv\Reader;
 
 class UmweltbundesamtStationLoader extends AbstractStationLoader
 {
-    const SOURCE_URL = 'https://www.env-it.de/stationen/public/download.do?event=euMetaStation';
+    const SOURCE_URL = 'https://www.umweltbundesamt.de/api/air_data/v2/meta/json?use=measure&lang=de';
 
-    /** @var Reader $csv */
-    protected $csv;
+    const FIELD_ID = 0;
+    const FIELD_STATION_CODE = 1;
+    const FIELD_TITLE = 2;
+    const FIELD_CITY = 3;
+    const FIELD_START_DATE = 5;
+    const FIELD_LONGITUDE = 7;
+    const FIELD_LATITUDE = 8;
+    const FIELD_STATE_CODE = 12;
+    const FIELD_STATE = 13;
+    const FIELD_AREA_TYPE = 15;
+    const FIELD_STATION_TYPE = 16;
 
     /** @var bool $update */
     protected $update = false;
+
+    /** @var array $ubaStationList */
+    protected $ubaStationList = [];
 
     public function process(callable $callback): StationLoaderInterface
     {
         /** @var EntityManager $em */
         $em = $this->registry->getManager();
 
-        foreach ($this->csv as $stationData) {
+        foreach ($this->ubaStationList as $stationData) {
             $callback();
 
-            if (!$stationData['station_code']) {
+            if (!array_key_exists(self::FIELD_STATION_CODE, $stationData) || !$stationData[self::FIELD_STATION_CODE]) {
                 continue;
-            } elseif (!$this->stationExists($stationData['station_code'], $this->existingStationList)) {
-                $station = $this->createStation($stationData);
-
-                $em->merge($station);
-
-                $this->newStationList[] = $station;
-            } elseif ($this->update === true) {
-                $station = $this->existingStationList[$stationData['station_code']];
-
-                $station = $this->mergeStation($station, $stationData);
             }
-        }
+
+            $stationCode = $stationData[self::FIELD_STATION_CODE];
+
+            if (!$this->stationExists($stationCode)) {
+                    $station = $this->createStation($stationData);
+
+                    $em->persist($station);
+
+                    $this->newStationList[] = $station;
+                } elseif ($this->update === true) {
+                    $station = $this->existingStationList[$stationCode];
+
+                    $station = $this->mergeStation($station, $stationData);
+                }
+            }
 
         $em->flush();
 
@@ -51,16 +68,15 @@ class UmweltbundesamtStationLoader extends AbstractStationLoader
     protected function mergeStation(Station $station, array $stationData): Station
     {
         $station
-            ->setTitle($stationData['station_name'])
+            ->setTitle($stationData[self::FIELD_TITLE])
             ->setProvider(UmweltbundesamtDeProvider::IDENTIFIER)
-            ->setStationCode($stationData['station_code'])
-            ->setLatitude(floatval($stationData['station_latitude_d']))
-            ->setLongitude(floatval($stationData['station_longitude_d']))
-            ->setFromDate($this->parseDate($stationData['station_start_date']))
-            ->setUntilDate(!empty($stationData['station_end_date']) ? $this->parseDate($stationData['station_end_date']) :null)
-            ->setAltitude(intval($stationData['station_altitude']))
-            ->setStationType(!empty($stationData['type_of_station']) ? $stationData['type_of_station'] : null)
-            ->setAreaType(!empty($stationData['station_type_of_area']) ? $stationData['station_type_of_area'] : null);
+            ->setStationCode($stationData[self::FIELD_STATION_CODE])
+            ->setLatitude((float) $stationData[self::FIELD_LATITUDE])
+            ->setLongitude((float) $stationData[self::FIELD_LONGITUDE])
+            ->setFromDate(new \DateTime($stationData[self::FIELD_START_DATE]))
+            ->setStationType($this->mapStationType($stationData[self::FIELD_STATION_TYPE]))
+            ->setAreaType($this->mapAreaType($stationData[self::FIELD_AREA_TYPE]))
+            ->setUbaStationId($stationData[self::FIELD_ID]);
 
         return $station;
     }
@@ -69,18 +85,14 @@ class UmweltbundesamtStationLoader extends AbstractStationLoader
     {
         $this->existingStationList = $this->getExistingStationList('uba_de');
 
-        $this->csv = $this->fetchStationList();
-
-        $this->csv
-            ->setDelimiter(';')
-            ->setHeaderOffset(1);
+        $this->fetchStationList();
 
         return $this;
     }
 
     public function count(): int
     {
-        return $this->csv ? $this->csv->count() : 0;
+        return count($this->ubaStationList);
     }
 
     public function setUpdate(bool $update = false): StationLoaderInterface
@@ -90,31 +102,22 @@ class UmweltbundesamtStationLoader extends AbstractStationLoader
         return $this;
     }
 
-    protected function fetchStationList(): Reader
+    protected function fetchStationList(): array
     {
         $curl = new Curl();
         $curl->get(self::SOURCE_URL);
 
-        $csv = Reader::createFromString(utf8_decode($curl->response));
+        $this->ubaStationList = (json_decode($curl->rawResponse, true))['stations'];
 
-        return $csv;
-    }
-
-    protected function parseStateCode(string $stationCode): string
-    {
-        return substr($stationCode, 2, 2);
-    }
-
-    protected function parseDate(string $dateString): \DateTime
-    {
-        sscanf($dateString,'%4d%2d%2d', $year, $month, $day);
-
-        return new \DateTime(sprintf('%d-%d-%d', $year, $month, $day));
+        return $this->ubaStationList;
     }
 
     protected function createStation(array $stationData): Station
     {
-        $station = new Station(floatval($stationData['station_latitude_d']), floatval($stationData['station_longitude_d']));
+        $latitude = (float) $stationData[self::FIELD_LATITUDE];
+        $longitude = (float) $stationData[self::FIELD_LONGITUDE];
+
+        $station = new Station($latitude, $longitude);
 
         $this->mergeStation($station, $stationData);
 
@@ -124,5 +127,24 @@ class UmweltbundesamtStationLoader extends AbstractStationLoader
     public function getExistingStationList(): array
     {
         return $this->registry->getRepository(Station::class)->findIndexedByProvider(UmweltbundesamtDeProvider::IDENTIFIER);
+    }
+
+    protected function mapAreaType(string $areaType): string
+    {
+        switch ($areaType) {
+            case 'vorstädtisch': return AreaType::SUBURBAN;
+            case 'städtisch': return AreaType::URBAN;
+            case 'ländlich': return AreaType::RURAL;
+        }
+    }
+
+    protected function mapStationType(string $stationType): string
+    {
+        switch ($stationType) {
+            case 'Hintergrund': return StationType::BACKGROUND;
+            case 'Verkehr': return StationType::TRAFFIC;
+            case 'Industrie': return StationType::INDUSTRIAL;
+            default: dump($stationType);
+        }
     }
 }
